Wordsearch Client and Server
This program creates a wordsearch with words defined from the client.
This works by having client program on one machine, and server program
on another machine, server runs the server program and allows access
through the defined socket, then the client enters the server IP, and
sends data to the server for the server to create the wordsearch.